requirements.txt
flask
qdrant-client
sentence-transformers
transformers
accelerate
python-dotenv
openai